requirements.txt
opencv-python>=4.8
tensorflow>=2.18
numpy>=1.24
matplotlib>=3.7
scikit-learn>=1.3
h5py>=3.9
pytube>=15
Pillow>=10
scipy>=1.11
